HiddenTambler/TagMo #6: Ability to edit Wolf Link heats
