+++ .gitignore
@@ -11,3 +11,4 @@
 /app/build
 /app/libs
 /app/app.iml
+/app/.externalNativeBuild
