+++ app/src/main/AndroidManifest.xml
@@ -22,8 +22,9 @@
             </intent-filter>
         </activity>
         <activity android:name=".NfcActivity_" android:launchMode="singleTask" />
-        <activity android:name=".EditorSSB_" />
-        <activity android:name=".HexViewerActivity_"></activity>
+        <activity android:name=".editors.EditorSSB_" />
+        <activity android:name=".editors.EditorTP_" />
+        <activity android:name=".HexViewerActivity_" />
     </application>
 
 </manifest>
