+++ app/src/main/res/layout/activity_editor_ssb.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.hiddenramblings.tagmo.EditorSSB">
+    tools:context="com.hiddenramblings.tagmo.editors.EditorSSB">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -16,7 +16,7 @@
             android:paddingLeft="@dimen/activity_horizontal_margin"
             android:paddingRight="@dimen/activity_horizontal_margin"
             android:paddingTop="@dimen/activity_vertical_margin"
-            tools:context="com.hiddenramblings.tagmo.EditorSSB"
+            tools:context="com.hiddenramblings.tagmo.editors.EditorSSB"
             android:layout_height="wrap_content"
             android:stretchColumns="1">
             <TableRow>
