+++ app/src/main/res/values/editor_tp_values.xml
@@ -1,0 +1,150 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="editor_tp_levels">
+        <item>0</item>
+        <item>1</item>
+        <item>2</item>
+        <item>3</item>
+        <item>4</item>
+        <item>5 (Run 1 completed)</item>
+        <item>6</item>
+        <item>7</item>
+        <item>8</item>
+        <item>9</item>
+        <item>10</item>
+        <item>11</item>
+        <item>12</item>
+        <item>13</item>
+        <item>14</item>
+        <item>15</item>
+        <item>16</item>
+        <item>17</item>
+        <item>18</item>
+        <item>19</item>
+        <item>20 (Run 2 completed)</item>
+        <item>21</item>
+        <item>22</item>
+        <item>23</item>
+        <item>24</item>
+        <item>25</item>
+        <item>26</item>
+        <item>27</item>
+        <item>28</item>
+        <item>29</item>
+        <item>30</item>
+        <item>31</item>
+        <item>32</item>
+        <item>33</item>
+        <item>34</item>
+        <item>35</item>
+        <item>36</item>
+        <item>37</item>
+        <item>38</item>
+        <item>39</item>
+        <item>40 (completed)</item>
+    </string-array>
+    <string-array name="editor_tp_hearts">
+        <item>0 (Dead!?)</item>
+
+        <item><![CDATA[0 + \u00BC]]></item>
+        <item><![CDATA[0 + \u00BD]]></item>
+        <item><![CDATA[0 + \u00BE]]></item>
+        
+        <item>1</item>
+        <item><![CDATA[1 + \u00BC]]></item>
+        <item><![CDATA[1 + \u00BD]]></item>
+        <item><![CDATA[1 + \u00BE]]></item>
+        
+        <item>2</item>
+        <item><![CDATA[2 + \u00BC]]></item>
+        <item><![CDATA[2 + \u00BD]]></item>
+        <item><![CDATA[2 + \u00BE]]></item>
+        
+        <item>3</item>
+        <item><![CDATA[3 + \u00BC]]></item>
+        <item><![CDATA[3 + \u00BD]]></item>
+        <item><![CDATA[3 + \u00BE]]></item>
+        
+        <item>4</item>
+        <item><![CDATA[4 + \u00BC]]></item>
+        <item><![CDATA[4 + \u00BD]]></item>
+        <item><![CDATA[4 + \u00BE]]></item>
+        
+        <item>5</item>
+        <item><![CDATA[5 + \u00BC]]></item>
+        <item><![CDATA[5 + \u00BD]]></item>
+        <item><![CDATA[5 + \u00BE]]></item>
+        
+        <item>6</item>
+        <item><![CDATA[6 + \u00BC]]></item>
+        <item><![CDATA[6 + \u00BD]]></item>
+        <item><![CDATA[6 + \u00BE]]></item>
+        
+        <item>7</item>
+        <item><![CDATA[7 + \u00BC]]></item>
+        <item><![CDATA[7 + \u00BD]]></item>
+        <item><![CDATA[7 + \u00BE]]></item>
+        
+        <item>8</item>
+        <item><![CDATA[8 + \u00BC]]></item>
+        <item><![CDATA[8 + \u00BD]]></item>
+        <item><![CDATA[8 + \u00BE]]></item>
+        
+        <item>9</item>
+        <item><![CDATA[9 + \u00BC]]></item>
+        <item><![CDATA[9 + \u00BD]]></item>
+        <item><![CDATA[9 + \u00BE]]></item>
+        
+        <item>10</item>
+        <item><![CDATA[10 + \u00BC]]></item>
+        <item><![CDATA[10 + \u00BD]]></item>
+        <item><![CDATA[10 + \u00BE]]></item>
+        
+        <item>11</item>
+        <item><![CDATA[11 + \u00BC]]></item>
+        <item><![CDATA[11 + \u00BD]]></item>
+        <item><![CDATA[11 + \u00BE]]></item>
+        
+        <item>12</item>
+        <item><![CDATA[12 + \u00BC]]></item>
+        <item><![CDATA[12 + \u00BD]]></item>
+        <item><![CDATA[12 + \u00BE]]></item>
+        
+        <item>13</item>
+        <item><![CDATA[13 + \u00BC]]></item>
+        <item><![CDATA[13 + \u00BD]]></item>
+        <item><![CDATA[13 + \u00BE]]></item>
+        
+        <item>14</item>
+        <item><![CDATA[14 + \u00BC]]></item>
+        <item><![CDATA[14 + \u00BD]]></item>
+        <item><![CDATA[14 + \u00BE]]></item>
+        
+        <item>15</item>
+        <item><![CDATA[15 + \u00BC]]></item>
+        <item><![CDATA[15 + \u00BD]]></item>
+        <item><![CDATA[15 + \u00BE]]></item>
+        
+        <item>16</item>
+        <item><![CDATA[16 + \u00BC]]></item>
+        <item><![CDATA[16 + \u00BD]]></item>
+        <item><![CDATA[16 + \u00BE]]></item>
+        
+        <item>17</item>
+        <item><![CDATA[17 + \u00BC]]></item>
+        <item><![CDATA[17 + \u00BD]]></item>
+        <item><![CDATA[17 + \u00BE]]></item>
+        
+        <item>18</item>
+        <item><![CDATA[18 + \u00BC]]></item>
+        <item><![CDATA[18 + \u00BD]]></item>
+        <item><![CDATA[18 + \u00BE]]></item>
+        
+        <item>19</item>
+        <item><![CDATA[19 + \u00BC]]></item>
+        <item><![CDATA[19 + \u00BD]]></item>
+        <item><![CDATA[19 + \u00BE]]></item>
+        
+        <item>20</item>
+    </string-array>
+</resources>
